backend/requirements.txt
# AgentFlow SDK
agentflow @ git+https://github.com/lupantech/AgentFlow.git

# Azure OpenAI
openai>=1.0.0
azure-identity>=1.15.0

# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
python-dateutil>=2.8.2
pytz>=2024.1

# ML & Data
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
